Renamed server/db.py to server/pebble_db.py
Another package named [db](https://pypi.python.org/pypi/db/0.1.1) is
located in the pypi. This might conflict with the app.
